+++ src/ff8/vibration.cpp
@@ -29,11 +29,10 @@
 #include <xxhash.h>
 
 char vibrateName[32] = "";
+ff8_menu_config_input_keymap menu_options_keymap_desc[11];
 
 int ff8_vibrate_capability(int port)
 {
-	bool ret = false;
-
 	if (trace_all) ffnx_trace("%s port=%d\n", __func__, port);
 
 	return nxVibrationEngine.canRumble();
@@ -242,11 +241,78 @@ int ff8_set_vibration(const uint8_t *data, int set, int intensity)
 	return ret;
 }
 
+ff8_draw_menu_sprite_texture_infos *alternate_pause_battle(int a1, ff8_draw_menu_sprite_texture_infos *draw_infos)
+{
+	if (nxVibrationEngine.canRumble())
+	{
+		uint16_t keyon = *(uint16_t *)(*(int *)ff8_externals.battle_menu_state + 0x14);
+		uint8_t port = 0;
+		int vibrate_state = ((int(*)(uint8_t))ff8_externals.vibration_get_is_enabled)(port);
+
+		if ((keyon & 0xF060) != 0)
+		{
+			vibrate_state = uint8_t(-(vibrate_state != 255));
+			((void(*)(uint8_t,int))ff8_externals.vibration_set_is_enabled)(port, vibrate_state);
+		}
+
+		*ff8_externals.is_alternative_pause_menu = 1;
+		int previous_value = *ff8_externals.pause_menu_option_state;
+		*ff8_externals.pause_menu_option_state = vibrate_state;
+
+		ff8_draw_menu_sprite_texture_infos *ret = ((ff8_draw_menu_sprite_texture_infos*(*)(int,ff8_draw_menu_sprite_texture_infos*))ff8_externals.battle_pause_window_sub_4CD350)(a1, draw_infos);
+
+		// Reset values to prevent any issues outside the pause menu
+		*ff8_externals.is_alternative_pause_menu = 0;
+		*ff8_externals.pause_menu_option_state = previous_value;
+
+		return ret;
+	}
+
+	return ((ff8_draw_menu_sprite_texture_infos*(*)(int,ff8_draw_menu_sprite_texture_infos*))ff8_externals.battle_pause_window_sub_4CD350)(a1, draw_infos);
+}
+
+ff8_draw_menu_sprite_texture_infos *battle_pause_show_quit_game_text(int a1, ff8_draw_menu_sprite_texture_infos *draw_infos, int x, int y, uint8_t *text_data, int a6)
+{
+	// Replace to "Vibration"
+	return ((ff8_draw_menu_sprite_texture_infos*(*)(int,ff8_draw_menu_sprite_texture_infos*,int,int))ff8_externals.sub_4A7210)(a1, draw_infos, x, y);
+}
+
 void vibration_init()
 {
-	replace_call(uint32_t(ff8_externals.check_game_is_paused) + 0x88, ff8_game_is_paused);
 	replace_function(ff8_externals.get_vibration_capability, ff8_vibrate_capability);
 	replace_function(ff8_externals.vibration_apply, apply_vibrate_calc);
 	replace_function(ff8_externals.vibration_clear_intensity, noop_a1);
 	ff8_set_vibration_replace_id = replace_function(ff8_externals.set_vibration, ff8_set_vibration);
+
+	// Replace pause menus
+	replace_call(uint32_t(ff8_externals.check_game_is_paused) + 0x88, ff8_game_is_paused);
+	replace_call(ff8_externals.battle_pause_sub_4CD140 + (JP_VERSION ? 0x1F1 : 0x225), alternate_pause_battle);
+	replace_call(ff8_externals.battle_pause_window_sub_4CD350 + 0x8D, battle_pause_show_quit_game_text);
+
+	// Relocate menu_options_keymap_desc to make room for a new input line in menu_options_desc which is located right before
+	memcpy(menu_options_keymap_desc, ff8_externals.menu_config_input_desc_keymap, sizeof(menu_options_keymap_desc));
+	patch_code_dword(uint32_t(ff8_externals.menu_callbacks[8].func) + 0x110, DWORD(menu_options_keymap_desc));
+	patch_code_dword(uint32_t(ff8_externals.menu_callbacks[8].func) + 0x12E, DWORD(menu_options_keymap_desc));
+	patch_code_dword(ff8_externals.menu_config_controller + 0x84, DWORD(menu_options_keymap_desc));
+	patch_code_dword(ff8_externals.menu_config_controller + 0x92, DWORD(menu_options_keymap_desc));
+	patch_code_dword(ff8_externals.menu_config_controller + 0xD7, DWORD(menu_options_keymap_desc) + 4);
+	patch_code_dword(ff8_externals.menu_config_controller + 0x109, DWORD(menu_options_keymap_desc) + 3);
+	patch_code_dword(ff8_externals.menu_config_controller + 0x155, DWORD(menu_options_keymap_desc) + 1);
+	patch_code_dword(ff8_externals.menu_config_controller + 0x53A, DWORD(menu_options_keymap_desc) + 3);
+	patch_code_dword(ff8_externals.menu_config_controller + 0x5B1, DWORD(menu_options_keymap_desc) + 3);
+	patch_code_dword(ff8_externals.menu_config_controller + 0x67B, DWORD(menu_options_keymap_desc));
+	patch_code_dword(ff8_externals.menu_config_controller + 0x7A0, DWORD(menu_options_keymap_desc) + 3);
+	patch_code_dword(ff8_externals.menu_config_render_submenu + 0x13, DWORD(menu_options_keymap_desc));
+
+	// Add Vibration configuration line in main menu
+	ff8_externals.menu_config_input_desc[9].text_id_name = 0x0; // "Vibration"
+	ff8_externals.menu_config_input_desc[9].text_id_value1 = 0x2; // "OFF"
+	ff8_externals.menu_config_input_desc[9].text_id_value2 = 0x1; // "ON"
+	ff8_externals.menu_config_input_desc[9].type = 2; // Toggle switch (Vibration option specific)
+	ff8_externals.menu_config_input_desc[9].mask = 0x40; // Bit mask in the savemap entry
+	ff8_externals.menu_config_input_desc[9].value = 0; // Always initialized to zero, will change at runtime
+	ff8_externals.menu_config_input_desc[9].callback_change_state = (void(*)())ff8_externals.sub_4C2FF0; // Update structures
+	// End of configuration lines
+	ff8_externals.menu_config_input_desc[10] = ff8_menu_config_input();
+	ff8_externals.menu_config_input_desc[10].text_id_name = -1;
 }